requirements.txt
fastapi
uvicorn
python-multipart
jinja2
sqlalchemy
alembic
python-jose
passlib
bcrypt
python-dotenv
pydantic
pillow
numpy
docker
aiosmtplib
aiofiles
authx
email-validator
python-docx
aiosqlite
starlette
typing-extensions
pathlib
uuid
python-jose[cryptography]
PyYAML